[issue-358] Move CLI tools to new model, change packaging in pyproject.toml to include only the src package
